requirements.txt
cv2>=3.4.2
google-cloud-vision
